Test Report Dir Bug Fix (#364)
A newly added feature was to create the directory for test reports if it didn't exist. The default if not specified location of the test report was "", which equated to the current directory. The check on "" existence was false which lead to the mkdir "" which is obvious going to fail. The solution is to verify that the artifactsDir isn't "" prior to that logic.

Signed-off-by: Ken Sipe <kensipe@gmail.com>