Improve error message for test assertions (#380)
* Bump controller-runtime to pick up a fake client fix.

We need [this commit](https://github.com/kubernetes-sigs/controller-runtime/commit/3a037ef46a00f5713b4d001804959f0ae812eadf)
for the subsequent changes which invoke List() on the fake client.

* Improve error message for error assertions and add test.

In the test cases for multiple objects we're using NewV1Pod, which returns an
actual Pod struct, rather than NewPod (which returns an Unstructured). This is
because the fake client is unable to List() pods which were supplied to builder
as Unstructureds. Instead it errors out with:

  item[0]: can't assign or convert unstructured.Unstructured into v1.Pod

Co-authored-by: Ken Sipe <kensipe@gmail.com>
Signed-off-by: Marcin Owsiany <porridge@redhat.com>