Drop [[toc]] until we find a better way.
GitHub automatically generates a ToC that seems to be hidden by default, but is
accessible using the "dotted hamburger" icon in upper right-hand corner. See
https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#headings

Ideally GitHub would provide a way to have an in-line auto-generated ToC or at
some point we can investigate ways to have it semi-automatically generated.

But for now having docs with this slightly hidden ToC is better than not having
them.

Signed-off-by: Marcin Owsiany <porridge@redhat.com>