chore: change namespace check and name determination
The main goal of this change is to determine the namespace name before
the test files are loaded. This way it will be available for template
expansion that will be added in an upcoming change.

The caveat is that determining of the name was established together with
deciding whether the namespace should be deleted after the test case
completes. However we want to avoid making cluster connections before
loading and validating test files.

To make it possible and simplify things a little at the same time, in
this change we no longer make a separate ns Get call. Instead we rely on
the response from the Create call. This has the additional advantage of
being atomic.

Signed-off-by: Marcin Owsiany <porridge@redhat.com>