chore: turn ns variable in Run() into a field
In an upcoming change, determineNamespace() will no longer be called by
Run()->setup() but instead by LoadTestSteps(). The resulting namespace
struct will need to be carried from LoadTestSteps() to Run() - both are
called by Harness. To keep the namespace encapsulated, we now save it as
a field inside Case.

As a result, some methods now access it from the receiver rather than an
explicit argument. Also, since determineNamespace no longer returns the
namespace, the control flow changed a bit to be easier to read, but is
functionally equivalent.

Signed-off-by: Marcin Owsiany <porridge@redhat.com>